Factor out `GameLog` as a component, and use the component in both places.
