+++ src/views/Game/ScoringEventThumbnail.tsx
